Add constant ID to static routes
refactor: Refactor remove method in EidPatternMap to accept generic id type

This change allows for greater flexibility in the types that can be used as identifiers when removing items from the map. The generic type constraint ensures that the id type implements the necessary traits for hashing and equality. This refactoring improves the code's maintainability and extensibility by making it more generic and reusable.