bpv7/tools: add CRC type argument to bundle create
The CRC types names are derived from the fields of the
hardy_bpv7::crc::CrcType struct.

The ArgCrcType enum value is translated into a u64 to be translated
into a CrcType using CrcType's From<u64> implementation.

Note that a current limitation of clap is that it will not print aliases
of the ValueEnum when --help is ran.
To compensate, I specified the aliases in the doc comment, which gets
printed when --help is ran.